requirements.txt
Django
django-social-auth
Pillow
markdown
django-registration
django-crispy-forms
MySQL-python
